requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
Django==2.2.5
django-debug-toolbar==2.0
django-filter==2.2.0
djangorestframework==3.10.3
factory-boy==2.12.0
Faker==2.0.2
importlib-metadata==0.23
inflection==0.3.1
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.2
pluggy==0.13.0
psycopg2-binary==2.8.3
py==1.8.0
pyparsing==2.4.2
pytest==5.1.2
pytest-django==3.5.1
pytest-factoryboy==2.0.3
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
wcwidth==0.1.7
zipp==0.6.0
